refactored to enforce improvedigital placement id for all auctions requests

+++ src/main/java/com/azerion/prebid/auction/model/GVastParams.java
@@ -10,7 +10,7 @@ import java.util.List;
 @Value
 public class GVastParams {
 
-    int placementId;
+    long placementId;
 
     Integer gdpr; // 1=gdpr applies
